+++ src/main/java/oripa/domain/cptool/PointSorter.java
@@ -1,0 +1,42 @@
+/**
+ * ORIPA - Origami Pattern Editor
+ * Copyright (C) 2013-     ORIPA OSS Project  https://github.com/oripa/oripa
+ * Copyright (C) 2005-2009 Jun Mitani         http://mitani.cs.tsukuba.ac.jp/
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package oripa.domain.cptool;
+
+import java.util.ArrayList;
+import java.util.Comparator;
+
+import javax.vecmath.Vector2d;
+
+import oripa.value.OriLine;
+
+/**
+ * @author OUCHI Koji
+ *
+ */
+class PointSorter {
+	public void sortPointsOnLine(final ArrayList<Vector2d> points, final OriLine line) {
+		boolean sortByX = Math.abs(line.p0.x - line.p1.x) > Math.abs(line.p0.y - line.p1.y);
+		if (sortByX) {
+			points.sort(Comparator.comparing(Vector2d::getX));
+		} else {
+			points.sort(Comparator.comparing(Vector2d::getY));
+		}
+	}
+
+}
